requirements.txt
flask
twilio
openai==0.28
python-dotenv
